requirements.txt
Flask==2.3.3
flask-cors==4.0.0
aiohttp==3.9.1
python-dotenv==1.0.1 